+++ core/src/test/kotlin/net/corda/core/contracts/TransactionGraphSearchTests.kt
@@ -1,7 +1,7 @@
 package net.corda.core.contracts
 
-import net.corda.core.contracts.testing.DummyContract
-import net.corda.core.contracts.testing.DummyState
+import net.corda.testing.contracts.DummyContract
+import net.corda.testing.contracts.DummyState
 import net.corda.core.crypto.newSecureRandom
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.WireTransaction