+++ core/src/test/kotlin/net/corda/core/contracts/clauses/VerifyClausesTests.kt
@@ -4,7 +4,7 @@ import net.corda.core.contracts.AuthenticatedObject
 import net.corda.core.contracts.CommandData
 import net.corda.core.contracts.ContractState
 import net.corda.core.contracts.TransactionForContract
-import net.corda.core.contracts.testing.DummyContract
+import net.corda.testing.contracts.DummyContract
 import net.corda.core.crypto.SecureHash
 import org.junit.Test
 import kotlin.test.assertFailsWith