+++ verifier/src/integration-test/kotlin/net/corda/verifier/GeneratedLedger.kt
@@ -2,7 +2,7 @@ package net.corda.verifier
 
 import net.corda.client.mock.*
 import net.corda.core.contracts.*
-import net.corda.core.contracts.testing.DummyContract
+import net.corda.testing.contracts.DummyContract
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.X509Utilities
 import net.corda.core.crypto.entropyToKeyPair